Renaming salt.modules.apt to salt.modules.apt_pkg to import python-apt
